requirements.txt
tqdm==4.66.1
pyexiftool==0.5.6
